requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.8
djangorestframework==3.7.3
enum==0.4.6
idna==2.6
pkg-resources==0.0.0
psutil==5.4.2
pytz==2017.3
requests==2.18.4
urllib3==1.22
